Fixed VS build.
git-svn-id: https://svnsrv.germanlloyd.org/svn/ExtFEMIO/trunk@195 389f8db8-dd28-11e4-b689-8b400cb78fb4

+++ src/fem/fem_cards_gnode.cpp
@@ -65,7 +65,7 @@ namespace dnvgl {
           const ::std::deque<int> *ODOF) :
           NODEX(*NODEX), NODENO(*NODENO),
           ODOF(ODOF->begin(), ODOF->end()) {
-          NDOF = ODOF->size();
+          NDOF = static_cast<long>(ODOF->size());
           ::std::sort(this->ODOF.begin(), this->ODOF.end());
         };
